Reworked SASS file structure
