+++ src/components/footer.js
@@ -1,7 +1,6 @@
 import React from 'react';
 // import {NavLink as RRNavLink} from 'react-router-dom';
 import logo from 'images/By Forte Primary (Black).png'
-import './footer.css';
 
 export class Footer extends React.Component {
   render() {
@@ -10,7 +9,7 @@ export class Footer extends React.Component {
         <hr/>
           <div className="row">
             <div className="col-sm">
-              <img className="" src={logo} alt="By Forte"/>
+              <img id="logo" src={logo} alt="By Forte"/>
               <br/>
               <p>Lorem ipsum dolor sit amet, consectetur adipiscing elit. Curabitur eget diam finibus, consectetur est vitae, porttitor sapien</p>
             </div>
