--- src/components/navbar/navbar.scss
@@ -1,19 +1,0 @@
-#wrapper {
-  padding-top: 2vh;
-  max-width: 850px;
-}
-
-img {
-  width: 70px;
-  height: 40px;
-  display: flex;
-  align-items: center;
-  flex-wrap: wrap;
-}
-
-
-.container {
-  width: auto;
-  max-width: 1000px;
-  padding: 0 15px;
-}
