--- src/scenes/about/about.js
@@ -1,13 +1,0 @@
-import React, {Component} from 'react';
-import './about.css';
-
-export class About extends Component {
-  render() {
-    return (<div className="container">
-      <h1>ABOUT</h1>
-      <p>Lorem ipsum dolor sit amet, consectetur adipiscing elit. Curabitur vel magna nec turpis tempor vehicula eu et leo. Nulla porta, quam in mollis aliquet, urna sapien pharetra tellus, ut scelerisque metus erat congue metus. Curabitur nec lacus accumsan, lobortis sapien et, elementum nunc. Curabitur vel purus sed nisi malesuada egestas.</p>
-      <p>Lorem ipsum dolor sit amet, consectetur adipiscing elit. Curabitur vel magna nec turpis tempor vehicula eu et leo. Nulla porta, quam in mollis aliquet, urna sapien pharetra tellus, ut scelerisque metus erat congue metus. Curabitur nec lacus accumsan, lobortis sapien et, elementum nunc. Curabitur vel purus sed nisi malesuada egestas.</p>
-      <p>Lorem ipsum dolor sit amet, consectetur adipiscing elit. Curabitur vel magna nec turpis tempor vehicula eu et leo. Nulla porta, quam in mollis aliquet, urna sapien pharetra tellus, ut scelerisque metus erat congue metus. Curabitur nec lacus accumsan, lobortis sapien et, elementum nunc. Curabitur vel purus sed nisi malesuada egestas.</p>
-    </div>)
-  }
-}
