+++ src/scenes/home.js
@@ -1,5 +1,5 @@
 import React, {Component} from 'react';
-import './home.css';
+import 'scss/scenes/home.css';
 
 export class Home extends Component {
   render() {
