+++ src/scss/components/_all.scss
@@ -1,0 +1,2 @@
+@import 'footer.scss';
+@import 'navbar.scss';
