+++ src/scss/components/_footer.scss
@@ -7,14 +7,13 @@
   min-height: 75px;
 }
 
-li {
+.footer > li {
   list-style: none;
 }
 
 .row {
   padding-top: 25px;
 }
-
 // Medium devices (tablets, 768px and up)
 @media (min-width: 576px) {
   .col-sm {
