+++ src/scss/components/_navbar.scss
@@ -1,0 +1,4 @@
+#wrapper {
+  padding-top: 2vh;
+  max-width: 850px;
+}
