+++ src/scss/main.scss
@@ -3,3 +3,7 @@
 
 // Partials
 @import 'scss/partials/typography';
+@import 'scss/partials/figures';
+
+// Components
+@import 'scss/components/all';
