+++ src/scss/modules/_all.scss
@@ -1,2 +1,1 @@
-@import 'colors';
 @import 'typography';
