+++ src/scss/partials/_base.scss
@@ -1,6 +1,6 @@
-@import url('https://fonts.googleapis.com/css?family=Lato|Oswald|Roboto');
+@import url('https://fonts.googleapis.com/css?family=Oswald|Roboto');
 $header-font: 'Oswald', sans-serif;
-$base-font: 'Lato', sans-serif;
+$base-font: 'Roboto', sans-serif;
 
 // Font weights
 $light: 100;
