+++ src/scss/partials/_figures.scss
@@ -1,0 +1,6 @@
+#logo {
+	height: 40px;
+	display: flex;
+	align-items: center;
+	flex-wrap: wrap;
+}
